@@ -1,29 +1,34 @@
 {
-  "name": "frantend",
+  "name": "inventory-billing",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.5.1",
-    "@mui/material": "^5.5.1",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.14.1",
+    "@mui/material": "^5.14.1",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "axios": "^0.26.1",
-    "bootstrap": "^5.1.3",
-    "caniuse-lite": "^1.0.30001481",
-    "formik": "^2.2.9",
-    "multiselect-react-dropdown": "^2.0.21",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
-    "react-google-charts": "^4.0.0",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "axios": "^1.4.0",
+    "bootstrap": "^5.3.0",
+    "formik": "^2.4.2",
+    "multiselect-react-dropdown": "^2.0.25",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.8.0",
+    "react-dom": "^18.2.0",
+    "react-google-charts": "^4.0.1",
+    "react-redux": "^8.1.1",
+    "react-router-dom": "^6.14.2",
+    "react-scripts": "5.0.1",
     "react-to-pdf": "^0.0.14",
+    "react-toastify": "^9.1.3",
+    "redux": "^4.2.1",
+    "redux-persist": "^6.0.0",
+    "redux-thunk": "^2.4.2",
     "web-vitals": "^2.1.4",
-    "yup": "^0.32.11"
+    "yup": "^1.2.0"
   },
   "scripts": {
     "start": "react-scripts start",
